+++ src/components/Sections/CommitAndResetSection/ResetButton.tsx
@@ -1,4 +1,4 @@
-import { resetSections } from '../store/store';
+import { resetSections } from '../../../store/store';
 
 export type ResetButtonProps = {
   disabled?: boolean;
@@ -10,6 +10,7 @@ export function ResetButton(props: ResetButtonProps) {
       className="focus:shadow-outline mx-2 my-6 rounded bg-white py-2 font-bold uppercase text-red-rhr hover:bg-red-200 focus:outline-none dark:bg-gray-500 dark:text-white dark:hover:bg-gray-700"
       type="button"
       onClick={() => resetSections()}
+      disabled={props.disabled}
     >
       Reset Form
     </button>
