+++ src/components/Sections/ConfigSection/index.ts
@@ -1,0 +1,1 @@
+export * from './ConfigSection';
